fix InvalidScssVariableException
